clean and add disposeWhere
