--- packages/app/src/app/overmind-graphql/effect.ts
@@ -1,231 +1,0 @@
-import * as withAbsintheSocket from '@absinthe/socket';
-import { GraphQLClient } from 'graphql-request';
-import {
-  Headers as HttpHeaders,
-  Options,
-} from 'graphql-request/dist/src/types';
-import { print } from 'graphql/language/printer';
-import { Socket as PhoenixSocket } from 'phoenix';
-
-export { default as gql } from 'graphql-tag';
-
-export interface Query<Result extends any, Payload extends any = void> {
-  (payload: Payload): Result;
-}
-
-interface NoPayloadSubscription<R> {
-  (action: (result: R) => void): void;
-  dispose(variables?: { [variables: string]: string | number | boolean }): void;
-}
-
-interface PayloadSubscription<P, R> {
-  (payload: P, action: (result: R) => void): void;
-  dispose(variables?: { [variables: string]: string | number | boolean }): void;
-}
-
-type Http = {
-  url: string;
-  headers?: () => HttpHeaders;
-  options?: Options;
-};
-
-type Ws = {
-  url: string;
-  params?: () => { [key: string]: string | number | boolean };
-};
-
-type Queries = {
-  queries?: {
-    [key: string]: (payload: any) => any;
-  };
-  mutations?: {
-    [key: string]: (payload: any) => any;
-  };
-  subscriptions?: {
-    [key: string]: (payload: any) => any;
-  };
-};
-
-export type Graphql<T extends Queries> = {
-  initialize(http: Http, ws?: Ws): void;
-} & {
-  queries: {
-    [N in keyof T['queries']]: T['queries'][N] extends (
-      payload: infer P
-    ) => infer R
-      ? P extends void
-        ? () => Promise<R>
-        : (payload: P) => Promise<R>
-      : never;
-  };
-  mutations: {
-    [N in keyof T['mutations']]: T['mutations'][N] extends (
-      payload: infer P
-    ) => infer R
-      ? P extends void
-        ? () => R
-        : (payload: P) => R
-      : never;
-  };
-  subscriptions: {
-    [N in keyof T['subscriptions']]: T['subscriptions'][N] extends (
-      payload: infer P
-    ) => infer R
-      ? P extends void
-        ? NoPayloadSubscription<R>
-        : PayloadSubscription<P, R>
-      : never;
-  };
-};
-
-function createError(message: string) {
-  throw new Error(`OVERMIND-GRAPHQL: ${message}`);
-}
-
-const _clients: { [url: string]: GraphQLClient } = {};
-const _wsClients: { [url: string]: PhoenixSocket } = {};
-const _subscriptions: {
-  [query: string]: { [variables: string]: () => void };
-} = {};
-
-export const graphql: <T extends Queries>(
-  queries: T
-) => Graphql<T> = queries => {
-  let _http: Http;
-  let _ws: Ws;
-
-  function getClient(): GraphQLClient | null {
-    if (_http) {
-      const headers = // eslint-disable-next-line
-        typeof _http.headers === 'function'
-          ? _http.headers()
-          : _http.options
-          ? _http.options.headers
-          : {};
-
-      if (_clients[_http.url]) {
-        _clients[_http.url].setHeaders(headers);
-      } else {
-        _clients[_http.url] = new GraphQLClient(_http.url, {
-          ..._http.options,
-          headers,
-        });
-      }
-
-      return _clients[_http.url];
-    }
-
-    return null;
-  }
-
-  function getWsClient(): PhoenixSocket | null {
-    if (_ws) {
-      if (!_wsClients[_ws.url]) {
-        _wsClients[_ws.url] = withAbsintheSocket.create(
-          new PhoenixSocket(_ws.url, {
-            params: _ws.params ? _ws.params() : null,
-          })
-        );
-      }
-
-      return _wsClients[_ws.url];
-    }
-
-    return null;
-  }
-
-  const evaluatedQueries = {
-    queries: Object.keys(queries.queries || {}).reduce((aggr, key) => {
-      aggr[key] = variables => {
-        const query = queries.queries[key] as any;
-        const client = getClient();
-
-        if (client) {
-          return client.request(print(query), variables);
-        }
-
-        throw createError(
-          'You are running a query, though there is no HTTP endpoint configured'
-        );
-      };
-      return aggr;
-    }, {}),
-    mutations: Object.keys(queries.mutations || {}).reduce((aggr, key) => {
-      aggr[key] = variables => {
-        const query = queries.mutations[key] as any;
-        const client = getClient();
-
-        if (client) {
-          return client.request(print(query), variables);
-        }
-
-        throw createError(
-          'You are running a query, though there is no HTTP endpoint configured'
-        );
-      };
-      return aggr;
-    }, {}),
-    subscriptions: Object.keys(queries.subscriptions || {}).reduce(
-      (aggr, key) => {
-        const query = queries.subscriptions[key] as any;
-        const queryString = print(query);
-
-        if (!_subscriptions[queryString]) {
-          _subscriptions[queryString] = {};
-        }
-
-        function subscription(arg1, arg2) {
-          const client = getWsClient();
-
-          if (client) {
-            const variables = arg2 ? arg1 : {};
-            const action = arg2 || arg1;
-            const notifier = withAbsintheSocket.send(client, {
-              operation: queryString,
-              variables,
-            });
-
-            const observer = withAbsintheSocket.observe(client, notifier, {
-              onResult: ({ data }) => {
-                action(data);
-              },
-            });
-
-            _subscriptions[queryString][JSON.stringify(variables)] = () =>
-              withAbsintheSocket.unobserve(client, notifier, observer);
-          }
-
-          throw createError('There is no ws client available for this query');
-        }
-
-        subscription.dispose = variables => {
-          if (variables && _subscriptions[queryString]) {
-            const subscriptionKey = JSON.stringify(variables);
-            _subscriptions[queryString][subscriptionKey]();
-            delete _subscriptions[queryString][subscriptionKey];
-          } else if (_subscriptions[queryString]) {
-            Object.keys(_subscriptions[queryString]).forEach(
-              subscriptionKey => {
-                _subscriptions[queryString][subscriptionKey]();
-                delete _subscriptions[queryString][subscriptionKey];
-              }
-            );
-          }
-        };
-
-        aggr[key] = subscription;
-
-        return aggr;
-      },
-      {}
-    ),
-  };
-
-  return {
-    initialize(http: Http, ws?: Ws) {
-      _http = http;
-      _ws = ws;
-    },
-    ...evaluatedQueries,
-  } as any;
-};
