+++ packages/app/src/app/overmind-graphql/index.ts
@@ -1,13 +1,11 @@
+import * as withAbsintheSocket from '@absinthe/socket';
 import { GraphQLClient } from 'graphql-request';
 import {
   Headers as HttpHeaders,
   Options,
 } from 'graphql-request/dist/src/types';
-import * as withAbsintheSocket from '@absinthe/socket';
-import { Socket as PhoenixSocket } from 'phoenix';
 import { print } from 'graphql/language/printer';
-import { IConfiguration } from 'overmind';
-import { ResolveState } from 'overmind/lib/internalTypes';
+import { Socket as PhoenixSocket } from 'phoenix';
 
 export { default as gql } from 'graphql-tag';
 
@@ -15,191 +13,218 @@ export interface Query<Result extends any, Payload extends any = void> {
   (payload: Payload): Result;
 }
 
+type Variable = string | number | boolean | null;
+
 interface NoPayloadSubscription<R> {
   (action: (result: R) => void): void;
-  dispose(variables?: { [variables: string]: string | number | boolean }): void;
+  dispose(): void;
+  disposeWhere(cb: (variables: { [variables: string]: Variable }) => boolean);
 }
 
 interface PayloadSubscription<P, R> {
   (payload: P, action: (result: R) => void): void;
-  dispose(variables?: { [variables: string]: string | number | boolean }): void;
+  dispose(): void;
+  disposeWhere(cb: (variables: { [variables: string]: Variable }) => boolean);
 }
 
-export function graphql<
-  C extends IConfiguration,
-  G extends {
-    endpoint: string;
-    headers?: (state: ResolveState<C['state']>) => HttpHeaders;
-    params?: (
-      state: ResolveState<C['state']>
-    ) => { [key: string]: string | number | boolean };
-    options?: Options;
-    queries?: {
-      [key: string]: (payload: any) => any;
-    };
-    mutations?: {
-      [key: string]: (payload: any) => any;
-    };
-    subscriptions?: {
-      [key: string]: (payload: any) => any;
-    };
-  }
->(
-  initialConfig: C,
-  options: G
-): {
-  state: C['state'];
-  effects: C['effects'] & {
-    queries: {
-      [N in keyof G['queries']]: G['queries'][N] extends (
-        payload: infer P
-      ) => infer R
-        ? P extends void
-          ? () => Promise<R>
-          : (payload: P) => Promise<R>
-        : never;
-    };
-    mutations: {
-      [N in keyof G['mutations']]: G['mutations'][N] extends (
-        payload: infer P
-      ) => infer R
-        ? P extends void
-          ? () => R
-          : (payload: P) => R
-        : never;
-    };
-    subscriptions: {
-      [N in keyof G['subscriptions']]: G['subscriptions'][N] extends (
-        payload: infer P
-      ) => infer R
-        ? P extends void
-          ? NoPayloadSubscription<R>
-          : PayloadSubscription<P, R>
-        : never;
-    };
+type Http = {
+  url: string;
+  headers?: () => HttpHeaders;
+  options?: Options;
+};
+
+type Ws = {
+  url: string;
+  params?: () => { [key: string]: string | number | boolean };
+};
+
+type Queries = {
+  queries?: {
+    [key: string]: (payload: any) => any;
   };
-  actions: C['actions'];
-} {
-  const config: any = {};
-  let _client: GraphQLClient;
-  let _subscriptionClient: any;
-  let _state;
-  const _subscriptions: {
-    [query: string]: { [variables: string]: () => void };
-  } = {};
-
-  if (config.effects && config.effects.queries) {
-    throw new Error(
-      `Overmind Graphql - There is already a "queries" effect on this configuration, please rename that to something else`
-    );
-  }
-
-  if (config.effects && config.effects.mutations) {
-    throw new Error(
-      `Overmind Graphql - There is already a "mutations" effect on this configuration, please rename that to something else`
-    );
-  }
-
-  if (config.effects && config.effects.subscriptions) {
-    throw new Error(
-      `Overmind Graphql - There is already a "subscriptions" effect on this configuration, please rename that to something else`
-    );
-  }
-
-  config.onInitialize = (context, payload) => {
-    _state = context.state;
-    _client = new GraphQLClient(options.endpoint, {
-      ...options.options,
-      headers:
-        typeof options.headers === 'function'
-          ? options.headers(context.state)
-          : options.options
-          ? options.options.headers
-          : {},
-    });
-    return (
-      initialConfig.onInitialize && initialConfig.onInitialize(context, payload)
-    );
+  mutations?: {
+    [key: string]: (payload: any) => any;
   };
-  config.state = initialConfig.state;
-  config.effects = Object.assign(initialConfig.effects || {}, {
-    queries: Object.keys(options.queries || {}).reduce((aggr, key) => {
+  subscriptions?: {
+    [key: string]: (payload: any) => any;
+  };
+};
+
+export type Graphql<T extends Queries> = {
+  initialize(http: Http, ws?: Ws): void;
+} & {
+  queries: {
+    [N in keyof T['queries']]: T['queries'][N] extends (
+      payload: infer P
+    ) => infer R
+      ? P extends void
+        ? () => Promise<R>
+        : (payload: P) => Promise<R>
+      : never;
+  };
+  mutations: {
+    [N in keyof T['mutations']]: T['mutations'][N] extends (
+      payload: infer P
+    ) => infer R
+      ? P extends void
+        ? () => R
+        : (payload: P) => R
+      : never;
+  };
+  subscriptions: {
+    [N in keyof T['subscriptions']]: T['subscriptions'][N] extends (
+      payload: infer P
+    ) => infer R
+      ? P extends void
+        ? NoPayloadSubscription<R>
+        : PayloadSubscription<P, R>
+      : never;
+  };
+};
+
+function createError(message: string) {
+  throw new Error(`OVERMIND-GRAPHQL: ${message}`);
+}
+
+const _clients: { [url: string]: GraphQLClient } = {};
+const _wsClients: { [url: string]: PhoenixSocket } = {};
+const _subscriptions: {
+  [query: string]: Array<{
+    variables: { [key: string]: Variable };
+    dispose: () => void;
+  }>;
+} = {};
+
+export const graphql: <T extends Queries>(
+  queries: T
+) => Graphql<T> = queries => {
+  let _http: Http;
+  let _ws: Ws;
+
+  function getClient(): GraphQLClient | null {
+    if (_http) {
+      const headers = // eslint-disable-next-line
+        typeof _http.headers === 'function'
+          ? _http.headers()
+          : _http.options
+          ? _http.options.headers
+          : {};
+
+      if (_clients[_http.url]) {
+        _clients[_http.url].setHeaders(headers);
+      } else {
+        _clients[_http.url] = new GraphQLClient(_http.url, {
+          ..._http.options,
+          headers,
+        });
+      }
+
+      return _clients[_http.url];
+    }
+
+    return null;
+  }
+
+  function getWsClient(): PhoenixSocket | null {
+    if (_ws) {
+      if (!_wsClients[_ws.url]) {
+        _wsClients[_ws.url] = withAbsintheSocket.create(
+          new PhoenixSocket(_ws.url, {
+            params: _ws.params ? _ws.params() : null,
+          })
+        );
+      }
+
+      return _wsClients[_ws.url];
+    }
+
+    return null;
+  }
+
+  const evaluatedQueries = {
+    queries: Object.keys(queries.queries || {}).reduce((aggr, key) => {
       aggr[key] = variables => {
-        const query = options.queries[key] as any;
+        const query = queries.queries[key] as any;
+        const client = getClient();
 
-        if (typeof options.headers === 'function') {
-          _client.setHeaders(options.headers(_state));
+        if (client) {
+          return client.request(print(query), variables);
         }
 
-        return _client.request(print(query), variables);
+        throw createError(
+          'You are running a query, though there is no HTTP endpoint configured'
+        );
       };
       return aggr;
     }, {}),
-    mutations: Object.keys(options.mutations || {}).reduce((aggr, key) => {
+    mutations: Object.keys(queries.mutations || {}).reduce((aggr, key) => {
       aggr[key] = variables => {
-        const query = options.mutations[key] as any;
+        const query = queries.mutations[key] as any;
+        const client = getClient();
 
-        if (typeof options.headers === 'function') {
-          _client.setHeaders(options.headers(_state));
+        if (client) {
+          return client.request(print(query), variables);
         }
 
-        return _client.request(print(query), variables);
+        throw createError(
+          'You are running a query, though there is no HTTP endpoint configured'
+        );
       };
       return aggr;
     }, {}),
-    subscriptions: Object.keys(options.subscriptions || {}).reduce(
+    subscriptions: Object.keys(queries.subscriptions || {}).reduce(
       (aggr, key) => {
-        const query = options.subscriptions[key] as any;
+        const query = queries.subscriptions[key] as any;
         const queryString = print(query);
 
         if (!_subscriptions[queryString]) {
-          _subscriptions[queryString] = {};
+          _subscriptions[queryString] = [];
         }
 
-        function subscription(variables, action) {
-          if (!_subscriptionClient) {
-            _subscriptionClient = withAbsintheSocket.create(
-              new PhoenixSocket('wss://codesandbox.test/graphql-socket', {
-                params: options.params ? options.params(_state) : null,
-              })
-            );
-          }
-          const notifier = withAbsintheSocket.send(_subscriptionClient, {
-            operation: queryString,
-            variables,
-          });
+        function subscription(arg1, arg2) {
+          const client = getWsClient();
 
-          const observer = withAbsintheSocket.observe(
-            _subscriptionClient,
-            notifier,
-            {
+          if (client) {
+            const variables = arg2 ? arg1 : {};
+            const action = arg2 || arg1;
+            const notifier = withAbsintheSocket.send(client, {
+              operation: queryString,
+              variables,
+            });
+
+            const observer = withAbsintheSocket.observe(client, notifier, {
               onResult: ({ data }) => {
                 action(data);
               },
-            }
-          );
+            });
 
-          _subscriptions[queryString][JSON.stringify(variables)] = () =>
-            withAbsintheSocket.unobserve(
-              _subscriptionClient,
-              notifier,
-              observer
-            );
+            _subscriptions[queryString].push({
+              variables,
+              dispose: () =>
+                withAbsintheSocket.unobserve(client, notifier, observer),
+            });
+          }
+
+          throw createError('There is no ws client available for this query');
         }
 
-        subscription.dispose = variables => {
-          if (variables && _subscriptions[queryString]) {
-            const subscriptionKey = JSON.stringify(variables);
-            _subscriptions[queryString][subscriptionKey]();
-            delete _subscriptions[queryString][subscriptionKey];
-          } else if (_subscriptions[queryString]) {
-            Object.keys(_subscriptions[queryString]).forEach(
-              subscriptionKey => {
-                _subscriptions[queryString][subscriptionKey]();
-                delete _subscriptions[queryString][subscriptionKey];
+        subscription.dispose = () => {
+          _subscriptions[queryString].forEach(sub => {
+            sub.dispose();
+          });
+          _subscriptions[queryString].length = 0;
+        };
+
+        subscription.disposeWhere = cb => {
+          _subscriptions[queryString] = _subscriptions[queryString].reduce(
+            (subAggr, sub) => {
+              if (cb(sub.variables)) {
+                return subAggr;
               }
-            );
-          }
+              return subAggr.concat(sub);
+            },
+            []
+          );
         };
 
         aggr[key] = subscription;
@@ -208,8 +233,13 @@ export function graphql<
       },
       {}
     ),
-  });
-  config.actions = initialConfig.actions;
+  };
 
-  return config as any;
-}
+  return {
+    initialize(http: Http, ws?: Ws) {
+      _http = http;
+      _ws = ws;
+    },
+    ...evaluatedQueries,
+  } as any;
+};
