+++ packages/app/src/app/overmind/effects/gql/collaborators/mutations.ts
@@ -14,7 +14,7 @@ import {
   RevokeSandboxInvitationMutation,
   RevokeSandboxInvitationMutationVariables,
 } from 'app/graphql/types';
-import { Query } from 'app/overmind-graphql/effect';
+import { Query } from 'app/overmind-graphql';
 import gql from 'graphql-tag';
 
 import { collaboratorFragment, sandboxInvitationFragment } from './fragments';
