+++ packages/app/src/app/overmind/effects/gql/collaborators/subscriptions.ts
@@ -12,7 +12,7 @@ import {
   OnInvitationRemovedSubscription,
   OnInvitationRemovedSubscriptionVariables,
 } from 'app/graphql/types';
-import { Query } from 'app/overmind-graphql/effect';
+import { Query } from 'app/overmind-graphql';
 import gql from 'graphql-tag';
 
 import { collaboratorFragment, sandboxInvitationFragment } from './fragments';
