+++ packages/app/src/app/overmind/effects/gql/index.ts
@@ -1,4 +1,4 @@
-import { graphql } from '../../../overmind-graphql/effect';
+import { graphql } from '../../../overmind-graphql';
 import * as mutations from './collaborators/mutations';
 import * as queries from './collaborators/queries';
 import * as subscriptions from './collaborators/subscriptions';
